+++ src/diffusers/models/__init__.py
@@ -35,8 +35,8 @@ if is_torch_available():
     _import_structure["unet_1d"] = ["UNet1DModel"]
     _import_structure["unet_2d"] = ["UNet2DModel"]
     _import_structure["unet_2d_condition"] = ["UNet2DConditionModel"]
-    _import_structure["unet_2d_condition_kandi3"] = ["Kandinsky3UNet"]
     _import_structure["unet_3d_condition"] = ["UNet3DConditionModel"]
+    _import_structure["unet_kandinsky3"] = ["Kandinsky3UNet"]
     _import_structure["unet_motion_model"] = ["MotionAdapter", "UNetMotionModel"]
     _import_structure["vq_model"] = ["VQModel"]
 
@@ -63,8 +63,8 @@ if TYPE_CHECKING or DIFFUSERS_SLOW_IMPORT:
         from .unet_1d import UNet1DModel
         from .unet_2d import UNet2DModel
         from .unet_2d_condition import UNet2DConditionModel
-        from .unet_2d_condition_kandi3 import Kandinsky3UNet
         from .unet_3d_condition import UNet3DConditionModel
+        from .unet_kandinsky3 import Kandinsky3UNet
         from .unet_motion_model import MotionAdapter, UNetMotionModel
         from .vq_model import VQModel
