+++ src/diffusers/models/unet_kandinsky3.py
